src/test/java/AppTest.java
import org.junit.*;
import static org.junit.Assert.*;


public class AppTest {

  @Test
  public void puzzleConverter_returnUpperCaseForbt_BT(){
    App app = new App();
    assertEquals("BT", app.puzzleConverter("bt"));
   }

   @Test
 		public void puzzle_returnDashForA_B_Dash_T() {
 		  App app = new App();
 		  assertEquals("B-T", app.puzzleConverter("BAT"));
 		  }

   @Test
    public void puzzle_returnDashForA_andE_B_T_() {
    	App app = new App();
    	assertEquals("B-T-", app.puzzleConverter("BATE"));
    	}

    @Test
      public void puzzle_returnDashForAEI_B__T_() {
        App app = new App();
        assertEquals("B--T-", app.puzzleConverter("BAITE"));
        }

     @Test
       public void puzzle_returnDashForAEIO_B___T_() {
        App app = new App();
        assertEquals("B---T-", app.puzzleConverter("BOAITE"));
        }

      @Test
       public void puzzle_returnDashForAEIOU_B___T__() {
         App app = new App();
         assertEquals("B---T--", app.puzzleConverter("BOAITUE"));
        }
 }
